Move host metrics gathering files into subdir
Bug: 550013241
