+++ base/cvd/cuttlefish/host/libs/metrics/conversion/BUILD.bazel
@@ -45,10 +45,10 @@ cf_cc_library(
     hdrs = ["host_conversion.h"],
     deps = [
         "//cuttlefish/common/libs/utils:host_info",
-        "//cuttlefish/host/libs/metrics:gce_environment",
-        "//cuttlefish/host/libs/metrics:github_environment",
-        "//cuttlefish/host/libs/metrics:host_metrics",
-        "//cuttlefish/host/libs/metrics:invoker",
+        "//cuttlefish/host/libs/metrics/host:gce_environment",
+        "//cuttlefish/host/libs/metrics/host:github_environment",
+        "//cuttlefish/host/libs/metrics/host:host_metrics",
+        "//cuttlefish/host/libs/metrics/host:invoker",
         "//external_proto:cf_gce_environment_cc_proto",
         "//external_proto:cf_github_actions_environment_cc_proto",
         "//external_proto:cf_host_cc_proto",
@@ -66,7 +66,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/metrics/conversion:guest_conversion",
         "//cuttlefish/host/libs/metrics:guest_metrics",
         "//cuttlefish/host/libs/metrics/conversion:host_conversion",
-        "//cuttlefish/host/libs/metrics:host_metrics",
+        "//cuttlefish/host/libs/metrics/host:host_metrics",
         "//external_proto:cf_log_cc_proto",
         "//external_proto:cf_metrics_event_v2_cc_proto",
         "@protobuf//:timestamp_cc_proto",
