+++ base/cvd/cuttlefish/host/libs/metrics/conversion/metrics_data.cc
@@ -25,7 +25,7 @@
 #include "cuttlefish/host/libs/metrics/conversion/host_conversion.h"
 #include "cuttlefish/host/libs/metrics/fetch_metrics.h"
 #include "cuttlefish/host/libs/metrics/guest_metrics.h"
-#include "cuttlefish/host/libs/metrics/host_metrics.h"
+#include "cuttlefish/host/libs/metrics/host/host_metrics.h"
 #include "external_proto/cf_log.pb.h"
 #include "external_proto/cf_metrics_event_v2.pb.h"
