+++ base/cvd/cuttlefish/host/libs/metrics/conversion/metrics_data.h
@@ -23,7 +23,7 @@
 
 #include "cuttlefish/host/libs/metrics/fetch_metrics.h"
 #include "cuttlefish/host/libs/metrics/guest_metrics.h"
-#include "cuttlefish/host/libs/metrics/host_metrics.h"
+#include "cuttlefish/host/libs/metrics/host/host_metrics.h"
 #include "external_proto/cf_log.pb.h"
 
 namespace cuttlefish {
