+++ base/cvd/cuttlefish/host/libs/metrics/host/gce_environment.h
