+++ base/cvd/cuttlefish/host/libs/metrics/host/host_metrics.h
@@ -20,9 +20,9 @@
 #include <variant>
 
 #include "cuttlefish/common/libs/utils/host_info.h"
-#include "cuttlefish/host/libs/metrics/gce_environment.h"
-#include "cuttlefish/host/libs/metrics/github_environment.h"
-#include "cuttlefish/host/libs/metrics/invoker.h"
+#include "cuttlefish/host/libs/metrics/host/gce_environment.h"
+#include "cuttlefish/host/libs/metrics/host/github_environment.h"
+#include "cuttlefish/host/libs/metrics/host/invoker.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
